requirements.txt
google-adk>=1.2.1
google-generativeai>=0.8.5
google-genai>=1.17.0
fastapi==0.115.12
uvicorn==0.34.2
httpx==0.28.1
pydantic==2.11.4
requests==2.32.3
psycopg2-binary==2.9.10
Deprecated==1.2.18

# A2A Protocol support
a2a>=0.0.3
sse-starlette>=2.0.0
python-multipart>=0.0.6
protobuf>=6.31.0,<7.0.0  # Match a2a SDK version